requirements.txt
pyyaml 
soundfile 
tqdm
keras-tuner
-e .